Extract all combining operators.
